python-server/requirements.txt
fastapi==0.74.0
uvicorn==0.16.0
marshmallow-sqlalchemy==0.26.1
